Refactor predicates, remove constructor
