+++ src/main/java/seedu/address/model/person/predicates/AppointmentOverlapsPredicate.java
@@ -1,8 +1,10 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
 
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.person.Appointment;
+import seedu.address.model.person.Person;
 
 /**
  * Tests that a {@code Person}'s {@code Appointment} overlaps with the query Appointment.
